+++ test/test_widgets/test_widgets.dart
@@ -29,3 +29,24 @@ class WidgetThree extends StatelessWidget {
   @override
   Widget build(BuildContext context) => Container();
 }
+
+class TestDashboard extends StatelessWidget {
+  final QRouter router;
+  TestDashboard(this.router);
+  @override
+  Widget build(BuildContext context) => Scaffold(
+        appBar: AppBar(
+          title: Text('Dashboard'),
+          centerTitle: true,
+        ),
+        body: Row(
+          children: [
+            Flexible(
+                child: Container(
+              child: Text('Sidebar'),
+            )),
+            Expanded(flex: 4, child: router)
+          ],
+        ),
+      );
+}